rooms: use TR2 scheduling; merge to trx
