+++ docs/tr1/CHANGELOG.md
@@ -50,6 +50,7 @@
 - fixed select level feature slot status not updated on save
 - fixed "Story So Far..." showing loading screens
 - fixed photo mode camera clipping through overlapping rooms
+- fixed matrix stack overflow crash when moving through overlapping or dome portals; TR1 now matches TR2 behavior (stable, with minor underside clipping) (#2685)
 
 ## [4.15.1](https://github.com/LostArtefacts/TRX/compare/tr1-4.15...tr1-4.15.1) - 2025-10-10
 - changed the examine dialog to be usable with non-puzzle items (#4009)
