+++ src/tr1/meson.build
@@ -110,7 +110,6 @@ sources = [
   'game/objects/setup.c',
   'game/option/option.c',
   'game/option/option_compass.c',
-  'game/room_draw.c',
   'game/savegame/savegame_bson.c',
   'game/savegame/savegame_legacy.c',
   'game/stats/common.c',
