--- src/tr2/game/room_draw.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <libtrx/game/rooms/draw.h>
