+++ src/tr2/meson.build
@@ -105,7 +105,6 @@ sources = [
   'game/objects/setup.c',
   'game/option/option.c',
   'game/option/option_compass.c',
-  'game/room_draw.c',
   'game/savegame/savegame_bson.c',
   'game/savegame/savegame_legacy.c',
   'game/stats.c',
